Moving corpus_class.py to corpus.py
