+++ allofplos/corpus/__init__.py
@@ -1,3 +1,3 @@
-from .corpus_class import Corpus
+from .corpus import Corpus
 
 from .plos_corpus import *
